requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
chromadb>=0.4.22
langchain>=0.1.0
langchain-community>=0.0.20
